requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dateutil==2.9.0
PyYAML==6.0.2
tqdm==4.66.5
spacy==3.7.4
vaderSentiment==3.3.2
gensim==4.3.2
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2